@@ -1,18 +1,17 @@
 name: mobx_architecture
 description: A Flutter package designed to aid in implementing a consistent, widespread architecture pattern, and providing a way to easily, clearly and elegantly bind MobX to the UI.
-version: 1.0.4
+version: 2.0.0-nullsafety.0
 homepage: https://tanzilzubair.github.io
 repository: https://github.com/tanzilzubair/mobx_architecture
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.17.0 <2.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  provider: ^4.3.2+3
-  mobx: ^1.2.1+4
+  provider: ^5.0.0
+  mobx: ^2.0.0-nullsafety.3
 
 dev_dependencies:
   flutter_test:
