@@ -1,5 +1,5 @@
 class SizingInfo {
-  const SizingInfo({this.height, this.width});
+  const SizingInfo({required this.height, required this.width});
   final double width;
   final double height;
 }
